Sync upload of images with comments, allowing users to post images alongside comments.
