+++ portfolio/src/main/java/com/google/sps/servlets/DataServlet.java
@@ -14,6 +14,18 @@
 
 package com.google.sps.servlets;
 
+import com.google.appengine.api.blobstore.BlobInfo;
+import com.google.appengine.api.blobstore.BlobInfoFactory;
+import com.google.appengine.api.blobstore.BlobKey;
+import com.google.appengine.api.blobstore.BlobstoreService;
+import com.google.appengine.api.blobstore.BlobstoreServiceFactory;
+import com.google.appengine.api.images.ImagesService;
+import com.google.appengine.api.images.ImagesServiceFactory;
+import com.google.appengine.api.images.ServingUrlOptions;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.Map;
 import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
@@ -39,6 +51,7 @@ public class DataServlet extends HttpServlet {
   public static final String COMMENT_DATE = "date";
   public static final String COMMENT_NAME = "Comment";
   public static final String COMMENT_MAX = "comment-max";
+  public static final String COMMENT_IMAGE = "imageUrl";
 
   // Default for max number of comments to show.
   final int COMMENT_MAX_DEFAULT = 5;
@@ -94,8 +107,9 @@ public class DataServlet extends HttpServlet {
       // Build the comment.
       String text = (String) entity.getProperty(COMMENT_TEXT);
       Date date = (Date) entity.getProperty(COMMENT_DATE);
+      String imageUrl = (String) entity.getProperty(COMMENT_IMAGE);
       long id = entity.getKey().getId();
-      Comment comment = new Comment(id, text, date);
+      Comment comment = new Comment(id, text, date, imageUrl);
       comments.add(comment);
 
       // Update count, and stop adding comments if comment max limit is reached.
@@ -126,8 +140,13 @@ public class DataServlet extends HttpServlet {
     String text = getParameter(request, "text-input", "");
     Date date = new Date();
 
+    // Get the URL of the image that the user uploaded to Blobstore.
+    // If no image was uploaded, this will simply be undefined and will be 
+    // handled on the frontend.
+    String imageUrl = getUploadedFileUrl(request, "image");
+
     // Create a comment entity from the Comment object.
-    Comment commentObject = new Comment(text, date);
+    Comment commentObject = new Comment(text, date, imageUrl);
     Entity commentEntity = commentObject.createCommentEntity();
 
     // Add the comment entity to the DatastoreService.
@@ -162,6 +181,47 @@ public class DataServlet extends HttpServlet {
     return value;
   }
 
+  /** 
+   * Returns a URL that points to the uploaded file, or null if the user 
+   * didn't upload a file.
+   */
+  private String getUploadedFileUrl(HttpServletRequest request, String formInputElementName) {
+    BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
+    Map<String, List<BlobKey>> blobs = blobstoreService.getUploads(request);
+    List<BlobKey> blobKeys = blobs.get("image");
+
+    // User submitted form without selecting a file, so we can't get a URL. (dev server)
+    if (blobKeys == null || blobKeys.isEmpty()) {
+      return null;
+    }
+
+    // Our form only contains a single file input, so get the first index.
+    BlobKey blobKey = blobKeys.get(0);
+
+    // User submitted form without selecting a file, so we can't get a URL. (live server)
+    BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
+    if (blobInfo.getSize() == 0) {
+      blobstoreService.delete(blobKey);
+      return null;
+    }
+
+    // We could check the validity of the file here, e.g. to make sure it's an image file
+    // https://stackoverflow.com/q/10779564/873165
+
+    // Use ImagesService to get a URL that points to the uploaded file.
+    ImagesService imagesService = ImagesServiceFactory.getImagesService();
+    ServingUrlOptions options = ServingUrlOptions.Builder.withBlobKey(blobKey);
+
+    // To support running in Google Cloud Shell with AppEngine's devserver, we must use the relative
+    // path to the image, rather than the path returned by imagesService which contains a host.
+    try {
+      URL url = new URL(imagesService.getServingUrl(options));
+      return url.getPath();
+    } catch (MalformedURLException e) {
+      return imagesService.getServingUrl(options);
+    }
+  }
+
   /**
    * Inner class for the Comments posted by users on the portfolio site.
    */
@@ -174,17 +234,19 @@ public class DataServlet extends HttpServlet {
     // The fields that hold the relevant comment data.
     private String text;
     private Date date;
+    private String imageUrl;
 
     // This constructor can create a Comment Entity (no ID required).
-    public Comment(String text, Date date) {
-      this(0, text, date);
+    public Comment(String text, Date date, String imageUrl) {
+      this(0, text, date, imageUrl);
     }
 
     // This constructor can accept a Comment Entity object (includes ID).
-    public Comment(long id, String text, Date date) {
+    public Comment(long id, String text, Date date, String imageUrl) {
       this.id = id;
       this.text = text;
       this.date = date;
+      this.imageUrl = imageUrl;
     }
 
     public Entity createCommentEntity() {
@@ -192,6 +254,7 @@ public class DataServlet extends HttpServlet {
       Entity commentEntity = new Entity(COMMENT_NAME);
       commentEntity.setProperty(COMMENT_TEXT, this.text);
       commentEntity.setProperty(COMMENT_DATE, this.date);
+      commentEntity.setProperty(COMMENT_IMAGE, this.imageUrl);
       return commentEntity;
     }
   }
