+++ portfolio/src/main/webapp/index.html
@@ -25,18 +25,16 @@
       <button id="delete-all-comments" onclick="deleteAllComments()">
         Delete All Comments
       </button>
-      <form action="/data" method="POST">
+      <form id="upload-form" class="hidden" method="POST" 
+        enctype="multipart/form-data">
         <textarea id="comment-box" name="text-input"
           oninput="checkSubmit('comment-box', 'comment-submit')"  
           minlength="1" placeholder="Join the discussion..."></textarea>
         <br>
-        <input id="comment-submit" type="submit" />
-      </form>
-      <form id="upload-form" class="hidden" method="POST" enctype="multipart/form-data">
         <p>Upload an image:</p>
         <input type="file" name="image">
-        <br/><br/>
-        <button>Submit</button>
+        <br>
+        <input id="comment-submit" type="submit" />
       </form>
       <form action="/data" method="POST">
         <label for="comment-max">Max Comments Shown (between 1 and 50):</label>
